Debugged receiveMsg method
The method now seems to work, at least with the test messages that I
have used until now. It seems to handle MIME and attachments ok, as
well as HTML. More testing needs to be done though.

Also removed a typo that crashed the edit_list page. Also will need to
somehow reconcile the asynch ajax call in collectMsgs with the need to
loop through a series of email addresses. Left a note to that effect in
the file.