Debugging complete of Plugin Methods
All the methods of the plugin have now been successfully debugged, but
work remains on the plugin webpages and command line scripts. Have
begun with confirmMsg.php, for which the debugging is quite far along,
and perhaps almost complete.

As to the other files, emailajax.php works well when validating an
email address or verifying POP credentials, but part concerned with the
collection of messages from email mailboxes is going to have to be
largely rewritten.