Finished debugging my ajax server
Added $isSecure property to class to flag a secure connection. The new
plan is that with an insecure connection we will continue to use of the
plugin, but we will be careful not to expose any of the submission
email data of the mailing lists to the browser, that is, we will not
expose the email address and POP credentials. That will happen in the
second release. For now we require a secure connection, disabling
activation if it is not there.

Debugged the ajax server to collect messages for collectMsgs.php. This
now works. But the ajax in collectMsgs.php still doesn't work. I have
left some of my debugging stuff in the file. This is the last file left
to debug. Everything else seems to work OK as designed.