Approaching Next Release
Have tested and debugged pipe script generated by plugin. Have
extensively used pipe to test message confirmation. Confirmation pages
now do not require a login. I’m not happy about this, but it seems
impossible to provide an administrator login that does not involve some
kind of kludge  and that does not look bad to the user.